Extract a component for inner navigation between top level datasources
